# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: HealthLog API
  version: 1.3.0
  description: |
    REST API for HealthLog — a personal health-tracking PWA covering weight,
    blood pressure, pulse, mood, medication compliance, blood glucose, and
    sleep. The same API powers the Next.js web client and the native iOS
    client.

    ## Authentication

    All non-public endpoints accept either a session cookie (`healthlog_session`,
    obtained via password login or passkey verification) or a long-lived Bearer
    token (`Authorization: Bearer hlk_...`) issued via `/api/tokens`. The two
    schemes are interchangeable on user endpoints — clients pick whichever
    suits them (cookies for the web app, Bearer for the iOS client and
    machine-to-machine integrations). Admin endpoints (`/api/admin/*`) accept
    only the session cookie; Bearer tokens never elevate to admin.

    The API uses an envelope response shape: `{ data, error }`. On success,
    `data` carries the payload and `error` is `null`. On failure, `data` is
    `null` and `error` is a human-readable message.

    ## Conventions

    - All timestamps are ISO 8601 (`date-time`) in UTC.
    - All numeric IDs are CUIDs (opaque strings).
    - Pagination uses `limit` + `offset` query parameters; responses include
      a `meta: { total, limit, offset }` block.
    - Rate limiting is applied per-IP or per-user on auth, ingest, export,
      AI, and bug-report endpoints. Exceeding the budget yields a `429`.

    ## Admin endpoints

...
servers:
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email/username + password
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginInput' }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponseEnvelope' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/doctor-report:
    post:
      tags: [DoctorReport]
      summary: Aggregated health data for doctor-report PDF
      operationId: getDoctorReport
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  minimum: 1
                  maximum: 365
                  default: 90
      responses:
        '200':
          description: Report data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/DoctorReportResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }

  # ─── Onboarding ───────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: 'hlk_<hex>'
      description: |
        Long-lived API token issued via `POST /api/tokens`. Accepted on all
        user endpoints (interchangeable with the session cookie) and required
        on external ingest endpoints (e.g. `/api/ingest/medication`). Tokens
        are prefixed `hlk_` and stored as SHA-256 hashes server-side. Bearer
        tokens never grant admin access — admin routes require the session
        cookie. Permission scopes are enforced from the token's `permissions`
        array.
    sessionCookie:
      type: apiKey
      in: cookie
      name: healthlog_session
      description: |
        HttpOnly session cookie set by `/api/auth/login` and the passkey
        login flow. Authenticates the web app and (optionally) the iOS
        client when used in a web view.
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    Unprocessable:
      description: Validation failure
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    ApiEnvelope:
      type: object
      description: Standard envelope. On success `error` is `null`; on failure `data` is `null` and `error` is a string.
      properties:
        data:
          description: Response payload (typed per-endpoint).
        error:
          type: [string, 'null']
      required: [data, error]
    ApiError:
      type: object
      properties:
        data:
          type: 'null'
        error:
          type: string
      required: [data, error]
...
    Gender:
      type: string
      enum: [MALE, FEMALE]
...
    LoginInput:
      type: object
      properties:
        email:
          type: string
          description: Email or username.
        password: { type: string }
      required: [email, password]
    LoginResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                user:
                  type: object
                  properties:
                    id: { type: string }
                    username: { type: string }
                  required: [id, username]
                token:
                  type: string
                  description: |
                    Long-lived `hlk_*` Bearer token. Only returned when the
                    request carries `X-Client-Type: native` or a User-Agent
                    starting with `HealthLog-iOS`. Use it for subsequent
                    requests as `Authorization: Bearer <token>`.
                tokenExpiresAt:
                  type: string
                  format: date-time
                  description: ISO timestamp at which the auto-issued token expires.
...
    DoctorReportResponse:
      type: object
      properties:
        period:
          type: object
          properties:
            days: { type: integer }
            since: { type: string, format: date-time }
        patient:
          type: object
          properties:
            username: { type: [string, 'null'] }
            dateOfBirth: { type: [string, 'null'], format: date-time }
            gender:
              oneOf:
                - $ref: '#/components/schemas/Gender'
                - type: 'null'
            heightCm: { type: [number, 'null'] }
        measurements:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              properties:
                value: { type: number }
                measuredAt: { type: string, format: date-time }
        stats:
          type: object
          additionalProperties:
            type: object
            properties:
              avg: { type: number }
              min: { type: number }
              max: { type: number }
              count: { type: integer }
              latest: { type: number }
        glucoseStats:
          type: object
          additionalProperties:
            type: object
            properties:
              avg: { type: number }
              min: { type: number }
              max: { type: number }
              count: { type: integer }
              latest: { type: number }
        glucoseRanges:
          type: object
          additionalProperties:
            type: object
            properties:
              min: { type: number }
              max: { type: number }
        glucoseUnit: { type: string }
        bmi: { type: [number, 'null'] }
        compliance:
          type: object
          additionalProperties:
            type: object
            properties:
              total: { type: integer }
              taken: { type: integer }
              skipped: { type: integer }
              missed: { type: integer }
        medications:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              dose: { type: string }
              schedules:
                type: array
                items:
                  type: object
                  properties:
                    windowStart: { type: string }
                    windowEnd: { type: string }
                    label: { type: [string, 'null'] }
        mood:
          type: [object, 'null']
          properties:
            avg: { type: number }
            min: { type: number }
            max: { type: number }
            count: { type: integer }
            distribution:
              type: object
              additionalProperties: { type: integer }

    # ─── Onboarding ────────────────────────────────────────────
...